cli/dm/docker/open-data-mesh-ws/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
psycopg2-binary>=2.9.9
jsonschema>=4.22.0
requests>=2.32.0
